Added PeakDataExtractor and lifted MS1 production upstream (commit 2 of 3)
* Moved RT-window/XIC/peak-detection/apex resolution out of CoelutionScorer.ScoreCandidate into a PeakDataExtractor producer seam; ScoreCandidate now extracts, runs the 21 calculators, and assembles the entry.
* Produced the MS1 precursor XIC + isotope envelope in the extractor and exposed them on the Detailed tier, so ms1_precursor_coelution / ms1_isotope_cosine became pure consumers; removed the now-dead MS1 machinery from OspreyScoringContext.
* Dropped dead per-candidate libCosine/top6Matches work. Byte-identical on Stellar+Astral (golden/resume/HPC-chain).

See ai/todos/active/TODO-20260620_ospreysharp_scoring_peakdata_tiers.md
